+++ module/WeaponMastery.js
@@ -268,19 +268,11 @@ export class WeaponMastery {
                             });
 
                             if (proceed) {
-                                const saveDC = 8 + abilityMod + workflow.actor.system.attributes.prof;
-                                const config = {undefined, ability: "con", target: saveDC};
-                                const dialog = {};
-                                const message = {data: {speaker: ChatMessage.implementation.getSpeaker({actor: targetToken.actor})}};
-                                let saveResult = await targetToken.actor.rollSavingThrow(config, dialog, message);
-
-                                if (!saveResult[0].isSuccess) {
-                                    await game.trazzm.socket.executeAsGM("toggleStatusEffect",
-                                        {actorUuid: targetToken.actor.uuid, statusId: 'prone', enabled: true});
-                                    ChatMessage.create({
-                                        content: `${targetToken.name} is knocked prone.`,
-                                        speaker: ChatMessage.getSpeaker({actor: workflow.actor})
-                                    });
+                                const activityId = `topple-${ability}`;
+                                const activity = macroItem.system.activities.find(a => a.identifier === activityId);
+                                if (activity) {
+                                    const targetUuids = [targetToken.document.uuid];
+                                    await MidiQOL.completeActivityUse(activity, { midiOptions: { targetUuids } });
                                 }
                             }
                         }